codigo/etapa01/requirements.txt
requests==2.31.0
networkx==3.2.1
matplotlib==3.8.2
pandas==2.1.4
numpy==1.25.2
seaborn==0.13.0
python-dotenv==1.0.0
tqdm==4.66.1
plotly==5.17.0